Add AlphaStable continuous distribution with special cases (Cauchy, Gaussian, Levy)
