+++ sympy/stats/crv_types.py
@@ -4730,3 +4730,141 @@ def WignerSemicircle(name, R):
     """
 
     return rv(name, WignerSemicircleDistribution, (R,))
+
+
+#-------------------------------------------------------------------------------
+# Alpha-Stable distribution ------------------------------------------------
+#-------------------------------------------------------------------------------
+
+class AlphaStableDistribution(SingleContinuousDistribution):
+    r"""
+    Represents an alpha-stable distribution.
+
+    Parameters
+    ==========
+
+    alpha : Real number, 0 < alpha <= 2
+        Stability parameter (controls tail heaviness).
+        When alpha < 2, the distribution has heavy tails.
+    beta : Real number, -1 <= beta <= 1
+        Skewness parameter. beta = 0 gives a symmetric distribution,
+        beta > 0 skews right, beta < 0 skews left.
+    scale : Positive real number
+        Scale parameter (analogous to standard deviation for normal distribution).
+    location : Real number
+        Location parameter (analogous to mean for normal distribution).
+
+    Examples
+    ========
+
+    >>> from sympy.stats import AlphaStable, density
+    >>> from sympy import Symbol
+    >>> x = Symbol('x', real=True)
+    >>> X = AlphaStable('X', 1, 0, 1, 0)
+    >>> density(X)(x)
+    1/(pi*(x**2 + 1))
+
+    References
+    ==========
+
+    .. [1] https://en.wikipedia.org/wiki/Stable_distribution
+    .. [2] Nolan, J. P. (2020). Univariate stable distributions.
+
+    """
+
+    _argnames = ('alpha', 'beta', 'scale', 'location')
+
+    @staticmethod
+    def check(alpha, beta, scale, location):
+        _value_check(alpha > 0, "Alpha must be positive")
+        _value_check(alpha <= 2, "Alpha must be less than or equal to 2")
+        _value_check(beta >= -1, "Beta must be >= -1")
+        _value_check(beta <= 1, "Beta must be <= 1")
+        _value_check(scale > 0, "Scale must be positive")
+
+    @property
+    def set(self):
+        return Interval(-oo, oo)
+
+    def pdf(self, x):
+        alpha, beta, scale, location = self.alpha, self.beta, self.scale, self.location
+        z = (x - location) / scale
+
+        # Special case: alpha = 2 (Gaussian)
+        if alpha == 2:
+            return exp(-z**2 / 2) / (scale * sqrt(2*pi))
+
+        # Special case: alpha = 1, beta = 0 (Cauchy)
+        if alpha == 1 and beta == 0:
+            return 1 / (pi * scale * (1 + z**2))
+
+        # Special case: alpha = 0.5, beta = 1 (Levy)
+        if alpha == Rational(1, 2) and beta == 1:
+            return Piecewise(
+                (sqrt(scale / (2*pi)) * exp(-scale / (2*z)) / z**Rational(3, 2), z > 0),
+                (0, True)
+            )
+
+        raise NotImplementedError(
+            "No closed-form PDF for alpha=%s, beta=%s. "
+            "Available special cases: alpha=2, (alpha=1, beta=0), (alpha=0.5, beta=1)." % (alpha, beta)
+        )
+
+    def _characteristic_function(self, t):
+        alpha, beta, scale, location = self.alpha, self.beta, self.scale, self.location
+
+        sign_t = Piecewise((1, t > 0), (-1, t < 0), (0, True))
+
+        if alpha == 1:
+            return exp(
+                I*t*location
+                - scale * Abs(t) * (1 + I*beta * (2/pi) * sign_t * log(Abs(t)))
+            )
+
+        return exp(
+            I*t*location
+            - (scale * Abs(t))**alpha * (1 - I*beta * sign_t * tan(pi*alpha/2))
+        )
+
+
+def AlphaStable(name, alpha, beta=0, scale=1, location=0):
+    r"""
+    Create an alpha-stable random variable.
+
+    Parameters
+    ==========
+
+    name : Symbol or string
+        Name of the random variable.
+    alpha : Real number, 0 < alpha <= 2
+        Stability parameter.
+    beta : Real number, -1 <= beta <= 1
+        Skewness parameter (default: 0).
+    scale : Positive real number
+        Scale parameter (default: 1).
+    location : Real number
+        Location parameter (default: 0).
+
+    Examples
+    ========
+
+    >>> from sympy.stats import AlphaStable, density
+    >>> from sympy import Symbol
+    >>> x = Symbol('x', real=True)
+    >>> X = AlphaStable('X', 1, 0, 1, 0)
+    >>> density(X)(x)
+    1/(pi*(x**2 + 1))
+
+    See Also
+    ========
+
+    sympy.stats.Normal
+    sympy.stats.Cauchy
+
+    References
+    ==========
+
+    .. [1] https://en.wikipedia.org/wiki/Stable_distribution
+
+    """
+    return rv(name, AlphaStableDistribution, (alpha, beta, scale, location))
